# backend/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Hive API",
    "description": "API documentation for the Hive news platform backend.",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Home",
        "description": "Returns a welcome message.",
        "responses": {
          "200": {
            "description": "Welcome message",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Welcome to Hive!"
                }
              }
            }
          }
        }
      }
    },
    "/api/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Registers a new user. Required fields: username, email, password. Optionally, role may be set by an admin.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User registration details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": { "type": "string" },
                "email": { "type": "string" },
                "password": { "type": "string" },
                "role": { "type": "string", "example": "regular" }
              },
              "required": ["username", "email", "password"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" },
                "user_id": { "type": "string" }
              }
            }
          },
          "400": {
            "description": "Missing required fields or user already exists"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "summary": "User login",
        "description": "Authenticates a user using username or email and password. Returns JWT tokens in HTTP-only cookies (and in JSON response in testing mode).",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User login credentials",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username_or_email": { "type": "string" },
                "password": { "type": "string" }
              },
              "required": ["username_or_email", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Login successful" },
                "access_token": { "type": "string" },
                "refresh_token": { "type": "string" }
              }
            }
          },
          "401": {
            "description": "User not found or invalid credentials"
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "summary": "User logout",
        "description": "Clears JWT cookies to log out the user.",
        "responses": {
          "200": {
            "description": "Logout successful",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Logged out successfully" }
              }
            }
          }
        }
      }
    },
    "/api/refresh": {
      "post": {
        "summary": "Refresh JWT tokens",
        "description": "Refreshes the access token using the refresh token from the HTTP-only cookie.",
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Token refreshed successfully" },
                "access_token": { "type": "string" },
                "refresh_token": { "type": "string" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid refresh token"
          }
        }
      }
    },
    "/api/articles": {
      "get": {
        "summary": "List Articles",
        "description": "Retrieves a paginated list of articles. Query parameters: page (default 1) and limit (default 2).",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (default 1)",
            "required": false,
            "type": "integer",
            "default": 1
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of articles per page (default 2)",
            "required": false,
            "type": "integer",
            "default": 2
          }
        ],
        "responses": {
          "200": {
            "description": "List of articles",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "article_id": { "type": "string" },
                  "title": { "type": "string" },
                  "content": { "type": "string" },
                  "author": { "type": "string" },
                  "created_at": { "type": "string" },
                  "updated_at": { "type": "string" }
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters"
          }
        }
      },
      "post": {
        "summary": "Create Article",
        "description": "Creates a new article. Only moderators and admins are authorized to create articles.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Article data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string" },
                "content": { "type": "string" }
              },
              "required": ["title", "content"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Article created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Article created successfully" },
                "article_id": { "type": "string" }
              }
            }
          },
          "400": {
            "description": "Missing title or content"
          },
          "403": {
            "description": "User not authorized to create articles"
          }
        }
      }
    },
    "/api/articles/{article_id}": {
      "get": {
        "summary": "Retrieve Article",
        "description": "Retrieves the details of a specific article by its ID.",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "description": "ID of the article",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Article details",
            "schema": {
              "type": "object",
              "properties": {
                "article_id": { "type": "string" },
                "title": { "type": "string" },
                "content": { "type": "string" },
                "author": { "type": "string" },
                "created_at": { "type": "string" },
                "updated_at": { "type": "string" }
              }
            }
          },
          "404": {
            "description": "Article not found"
          }
        }
      },
      "put": {
        "summary": "Update Article",
        "description": "Updates an existing article. Requires at least one field (title or content) to update.",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "description": "ID of the article to update",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Fields to update",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string" },
                "content": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Article updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Article updated successfully" }
              }
            }
          },
          "400": {
            "description": "No update data provided"
          },
          "404": {
            "description": "Article not found"
          }
        }
      },
      "delete": {
        "summary": "Delete Article",
        "description": "Deletes an article. Only authorized users (author, moderator, or admin) can delete an article.",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "description": "ID of the article to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Article deleted successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "Article deleted successfully" }
              }
            }
          },
          "404": {
            "description": "Article not found"
          },
          "403": {
            "description": "User not authorized to delete this article"
          }
        }
      }
    },
    "/api/articles/search": {
      "get": {
        "summary": "Search Articles",
        "description": "Searches for articles based on a query parameter.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "Search keyword",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "article_id": { "type": "string" },
                  "title": { "type": "string" },
                  "content": { "type": "string" },
                  "author": { "type": "string" },
                  "created_at": { "type": "string" },
                  "updated_at": { "type": "string" }
                }
              }
            }
          },
          "400": {
            "description": "Missing search query parameter"
          }
        }
      }
    }
  }
}